moves code for fetching reference data to hooks
